multi: wrap math.MaxInt64 in int64() for 32-bit cross-compilation
We must wrap math.MaxInt64 calls in int64() for rpi/arm/32-bit support, otherwise cross-compilation yields this error:
```
cmd/tarocli/addrs.go:144:9: cannot use math.MaxInt64 (untyped int constant 9223372036854775807) as int value in assignment (overflows)
```

+++ cmd/tarocli/addrs.go
@@ -141,7 +141,8 @@ func queryAddr(ctx *cli.Context) error {
 		start = startTime.Unix()
 	}
 
-	end := math.MaxInt64
+	// Wrap with int64() so math.MaxInt64 will cross-compile on 32-bit arch.
+	end := int64(math.MaxInt64) 
 	if ctx.IsSet(createdBeforeName) {
 		endOffset, err := time.ParseDuration(ctx.String(createdBeforeName))
 		if err != nil {
